feat(referencefile_mod): automatically delete referencefile when the last referencefile_mod connection is deleted
